+++ util/src/main/java/net/automatalib/util/automaton/equivalence/DeterministicEquivalenceTest.java
@@ -37,16 +37,13 @@ public final class DeterministicEquivalenceTest {
         // prevent instantiation
     }
 
+    @SuppressWarnings("PMD.UnnecessaryCast") // we want to cast to long, to prevent overflows
     public static <I, S, T, SP, TP, S2, T2, SP2, TP2> @Nullable Word<I> findSeparatingWord(
             UniversalDeterministicAutomaton<S, I, T, SP, TP> reference,
             UniversalDeterministicAutomaton<S2, I, T2, SP2, TP2> other,
             Collection<? extends I> inputs) {
         int refSize = reference.size();
-        int totalStates = refSize * other.size();
-
-        if (totalStates > MAP_THRESHOLD) {
-            return findSeparatingWordLarge(reference, other, inputs);
-        }
+        int otherSize = other.size();
 
         S refInit = reference.getInitialState();
         S2 otherInit = other.getInitialState();
@@ -62,32 +59,36 @@ public final class DeterministicEquivalenceTest {
             return Word.epsilon();
         }
 
-        Queue<StatePair<S, S2>> bfsQueue = new ArrayDeque<>();
-        bfsQueue.add(new StatePair<>(refInit, otherInit));
-
         StateIDs<S> refStateIds = reference.stateIDs();
         StateIDs<S2> otherStateIds = other.stateIDs();
-
-        StatePair<S, S2> currPair;
-        int lastId = otherStateIds.getStateId(otherInit) * refSize + refStateIds.getStateId(refInit);
-
-        @SuppressWarnings("unchecked")
-        Pred<I>[] preds = new Pred[totalStates];
-        preds[lastId] = new Pred<>();
 
         int currDepth = 0;
         int inCurrDepth = 1;
         int inNextDepth = 0;
 
         I lastSym = null;
+        Pred<I> lastPred = new Pred<>();
+        Registry<I> reg;
+
+        if ((long) refSize * (long) otherSize > MAP_THRESHOLD) {
+            reg = new MapRegistry<>(refSize);
+        } else {
+            reg = new ArrayRegistry<>(refSize, otherSize);
+        }
+
+        reg.putPred(refStateIds.getStateId(refInit), otherStateIds.getStateId(otherInit), lastPred);
+
+        Queue<StatePair<S, S2>> bfsQueue = new ArrayDeque<>();
+        bfsQueue.add(new StatePair<>(refInit, otherInit));
+        StatePair<S, S2> currPair;
 
         bfs:
         while ((currPair = bfsQueue.poll()) != null) {
             S refState = currPair.ref;
             S2 otherState = currPair.other;
 
-            int currId = otherStateIds.getStateId(otherState) * refSize + refStateIds.getStateId(refState);
-            lastId = currId;
+            lastPred = reg.getPred(refStateIds.getStateId(refState), otherStateIds.getStateId(otherState));
+            assert lastPred != null;
 
             for (I in : inputs) {
                 lastSym = in;
@@ -111,9 +112,10 @@ public final class DeterministicEquivalenceTest {
                 S refSucc = reference.getSuccessor(refTrans);
                 S2 otherSucc = other.getSuccessor(otherTrans);
 
-                int succId = otherStateIds.getStateId(otherSucc) * refSize + refStateIds.getStateId(refSucc);
+                int refId = refStateIds.getStateId(refSucc);
+                int otherId = otherStateIds.getStateId(otherSucc);
 
-                if (preds[succId] == null) {
+                if (reg.getPred(refId, otherId) == null) {
                     refStateProp = reference.getStateProperty(refSucc);
                     otherStateProp = other.getStateProperty(otherSucc);
 
@@ -121,7 +123,7 @@ public final class DeterministicEquivalenceTest {
                         break bfs;
                     }
 
-                    preds[succId] = new Pred<>(currId, in);
+                    reg.putPred(refId, otherId, new Pred<>(lastPred, in));
                     bfsQueue.add(new StatePair<>(refSucc, otherSucc));
                     inNextDepth++;
                 }
@@ -146,125 +148,10 @@ public final class DeterministicEquivalenceTest {
         int index = currDepth;
         sep.setSymbol(index--, lastSym);
 
-        Pred<I> pred = preds[lastId];
-        while (pred.id >= 0) {
+        Pred<I> pred = lastPred;
+        while (pred.prev != null) {
             sep.setSymbol(index--, pred.symbol);
-            pred = preds[pred.id];
-        }
-
-        return sep.toWord();
-    }
-
-    public static <I, S, T, SP, TP, S2, T2, SP2, TP2> @Nullable Word<I> findSeparatingWordLarge(
-            UniversalDeterministicAutomaton<S, I, T, SP, TP> reference,
-            UniversalDeterministicAutomaton<S2, I, T2, SP2, TP2> other,
-            Collection<? extends I> inputs) {
-        S refInit = reference.getInitialState();
-        S2 otherInit = other.getInitialState();
-
-        if (refInit == null || otherInit == null) {
-            return refInit == null && otherInit == null ? null : Word.epsilon();
-        }
-
-        SP refStateProp = reference.getStateProperty(refInit);
-        SP2 otherStateProp = other.getStateProperty(otherInit);
-
-        if (!Objects.equals(refStateProp, otherStateProp)) {
-            return Word.epsilon();
-        }
-
-        Queue<StatePair<S, S2>> bfsQueue = new ArrayDeque<>();
-        bfsQueue.add(new StatePair<>(refInit, otherInit));
-
-        int refSize = reference.size();
-
-        StateIDs<S> refStateIds = reference.stateIDs();
-        StateIDs<S2> otherStateIds = other.stateIDs();
-
-        StatePair<S, S2> currPair;
-        int lastId = otherStateIds.getStateId(otherInit) * refSize + refStateIds.getStateId(refInit);
-
-        //TIntObjectMap<Pred<I>> preds = new TIntObjectHashMap<>();
-        Map<Integer, Pred<I>> preds = new HashMap<>(); // TODO: replace by primitive specialization
-        preds.put(lastId, new Pred<>());
-
-        int currDepth = 0;
-        int inCurrDepth = 1;
-        int inNextDepth = 0;
-
-        I lastSym = null;
-
-        bfs:
-        while ((currPair = bfsQueue.poll()) != null) {
-            S refState = currPair.ref;
-            S2 otherState = currPair.other;
-
-            int currId = otherStateIds.getStateId(otherState) * refSize + refStateIds.getStateId(refState);
-            lastId = currId;
-
-            for (I in : inputs) {
-                lastSym = in;
-                T refTrans = reference.getTransition(refState, in);
-                T2 otherTrans = other.getTransition(otherState, in);
-
-                if (refTrans == null || otherTrans == null) {
-                    if (refTrans == null && otherTrans == null) {
-                        continue;
-                    } else {
-                        break bfs;
-                    }
-                }
-
-                TP refProp = reference.getTransitionProperty(refTrans);
-                TP2 otherProp = other.getTransitionProperty(otherTrans);
-                if (!Objects.equals(refProp, otherProp)) {
-                    break bfs;
-                }
-
-                S refSucc = reference.getSuccessor(refTrans);
-                S2 otherSucc = other.getSuccessor(otherTrans);
-
-                int succId = otherStateIds.getStateId(otherSucc) * refSize + refStateIds.getStateId(refSucc);
-
-                if (preds.get(succId) == null) {
-                    refStateProp = reference.getStateProperty(refSucc);
-                    otherStateProp = other.getStateProperty(otherSucc);
-
-                    if (!Objects.equals(refStateProp, otherStateProp)) {
-                        break bfs;
-                    }
-
-                    preds.put(succId, new Pred<>(currId, in));
-                    bfsQueue.add(new StatePair<>(refSucc, otherSucc));
-                    inNextDepth++;
-                }
-            }
-
-            lastSym = null;
-
-            // Next level in BFS reached
-            if (--inCurrDepth == 0) {
-                inCurrDepth = inNextDepth;
-                inNextDepth = 0;
-                currDepth++;
-            }
-        }
-
-        if (lastSym == null) {
-            return null;
-        }
-
-        @SuppressWarnings("nullness") // we make sure to set each index to a value of type I
-        WordBuilder<I> sep = new WordBuilder<>(null, currDepth + 1);
-        int index = currDepth;
-        sep.setSymbol(index--, lastSym);
-
-        Pred<I> pred = preds.get(lastId);
-        assert pred != null;
-        while (pred.id >= 0) {
-            sep.setSymbol(index--, pred.symbol);
-            pred = preds.get(pred.id);
-            assert pred != null;
+            pred = pred.prev;
         }
 
         return sep.toWord();
@@ -283,18 +170,75 @@ public final class DeterministicEquivalenceTest {
 
     private static final class Pred<I> {
 
-        public final int id;
-        public final I symbol;
+        public final @Nullable Pred<I> prev;
+        public final @Nullable I symbol;
 
-        @SuppressWarnings("nullness") // we check this special element using its id value
         Pred() {
-            this.id = -1;
+            this.prev = null;
             this.symbol = null;
         }
 
-        Pred(int id, I input) {
-            this.id = id;
+        Pred(Pred<I> prev, I input) {
+            this.prev = prev;
             this.symbol = input;
+        }
+    }
+
+    private interface Registry<I> {
+
+        @Nullable Pred<I> getPred(int id1, int id2);
+
+        void putPred(int id1, int id2, Pred<I> pred);
+    }
+
+    private static class ArrayRegistry<I> implements Registry<I> {
+
+        final Pred<I>[] preds;
+        final int size1;
+
+        @SuppressWarnings("unchecked")
+        ArrayRegistry(int size1, int size2) {
+            this.preds = new Pred[size1 * size2];
+            this.size1 = size1;
+        }
+
+        @Override
+        public @Nullable Pred<I> getPred(int id1, int id2) {
+            return preds[computeIndex(id1, id2, size1)];
+        }
+
+        @Override
+        public void putPred(int id1, int id2, Pred<I> pred) {
+            preds[computeIndex(id1, id2, size1)] = pred;
+        }
+
+        private int computeIndex(int id1, int id2, int size1) {
+            return id1 * size1 + id2;
+        }
+    }
+
+    private static class MapRegistry<I> implements Registry<I> {
+
+        final Map<Long, Pred<I>> preds;
+        final int size1;
+
+        MapRegistry(int size1) {
+            this.preds = new HashMap<>();
+            this.size1 = size1;
+        }
+
+        @Override
+        public @Nullable Pred<I> getPred(int id1, int id2) {
+            return preds.get(computeIndex(id1, id2, size1));
+        }
+
+        @Override
+        public void putPred(int id1, int id2, Pred<I> pred) {
+            preds.put(computeIndex(id1, id2, size1), pred);
+        }
+
+        private long computeIndex(long id1, long id2, long size1) {
+            return id1 * size1 + id2;
         }
     }
 }